improve the dtype selection

+++ mace/calculators/foundations_models.py
@@ -51,9 +51,7 @@ def mace_mp(
 
     device = device or ("cuda" if torch.cuda.is_available() else "cpu")
 
-    return MACECalculator(
-        model_path, device=device, default_dtype=default_dtype, **kwargs
-    )
+    return MACECalculator(model_path, device=device, **kwargs)
 
 
 def mace_anicc(
